TypeTreesIO

Documentation for TypeTreesIO.

TypeTreesIO.TypeTreeIO

TypeTreesIO.width_by_depth

TypeTreesIO.TypeTreeIO — Type

TypeTreeIO() → io

Create an IO object to which you can print type objects or natural signatures. Afterwards, io.tree will contain a tree representation of the printed type.

Examples

julia> io = TypeTreeIO();

julia> print(io, Tuple{Int,Float64});

julia> io.tree.body.name
"Tuple"

julia> io.tree.body.children[1].name
"Int64"

julia> String(take!(io))
"Tuple{Int64, Float64}"

Extended help

In addition to printing a type directly to an io::TypeTreeIO, you can also assemble it manually if you follow a few precautions:

- any `where` statement must be printed as `print(io, " where ")` or
  `print(io, " where {")`. The `where` string argument may not have any
  additional characters. Note the bracketing spaces.

TypeTreesIO.width_by_depth — Method

width_by_depth(node) → wd, wtrunc

Compute the number of characters wd[depth] needed to print at each depth within the tree. Also compute the number of additional characters wtrunc[depth] needed if one truncates the tree at depth.